% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspi.R
\docType{package}
\name{aspi}
\alias{aspi}
\alias{aspi-package}
\title{ASPI: Analysis of Symmetry in Parasitic Infections}
\description{
Tools for the analysis and visualization of bilateral asymmetry in
parasitic infections.
}
\section{Functions}{

\itemize{
\item{\code{g.test}}
\item{\code{eb.test}}
\item{\code{plotHistogram}}
\item{\code{plotVolcano}}
}
}

\section{Data}{

\itemize{
\item{\code{diplostomum_eyes_excl_lenses}}
\item{\code{diplostomum_lenses}}
\item{\code{simulated_symmetrical_infection}}
\item{\code{simulated_left_bias_homogeneous_proportions}}
\item{\code{simulated_left_bias_heterogeneous_proportions}}
\item{\code{simulated_asymmetry_inconsistent_bias}}
}
}
\author{
Matt Wayland, \email{mw283@cam.ac.uk}
}

