% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diplostomum_eyes_excl_lenses}
\alias{diplostomum_eyes_excl_lenses}
\title{Numbers of Diplostomum metacercariae recorded from the eyes (excluding lenses) 
of each of 50 ruffe.}
\format{A data frame with 50 rows and 2 variables (row names are host IDs):
\describe{
 \item{left}{number of metacercariae in left eye}
 \item{right}{number of metacercariae in the right eye}
 }}
\usage{
diplostomum_eyes_excl_lenses
}
\description{
A dataset containing the numbers of Diplostomum metacercariae recovered from 
the eyes (excluding lenses) of each of 50 ruffe, Gymnocephalus cernuus from 
Llyn Tegid, a mesotrophic lake in north Wales.
}
\keyword{datasets}

