% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diplostomum_lenses}
\alias{diplostomum_lenses}
\title{Numbers of Diplostomum metacercariae recorded from the lenses of the eyes
of each of 50 ruffe.}
\format{A data frame with 50 rows and 2 variables (row names are host IDs):
\describe{
 \item{left}{number of metacercariae in lens of left eye}
 \item{right}{number of metacercariae in lens of right eye}
 }}
\usage{
diplostomum_lenses
}
\description{
A dataset containing the numbers of Diplostomum metacercariae recovered from 
the lenses of the eyes of each of 50 ruffe, Gymnocephalus cernuus from
Llyn Tegid, a mesotrophic lake in north Wales.
}
\keyword{datasets}

