% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_asymmetry_inconsistent_bias}
\alias{simulated_asymmetry_inconsistent_bias}
\title{Simulated data showing bilateral asymmetry with insconsistent bias}
\format{A data frame with 10 rows and 2 variables:
\describe{
 \item{left}{number of parasites on the left side}
 \item{right}{number of parasites on the right side}
 }}
\usage{
simulated_asymmetry_inconsistent_bias
}
\description{
Simulated data for 10 hosts, providing an example of bilateral asymmetry with inconsistent bias.
}
\keyword{datasets}

