% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_left_bias_heterogeneous_proportions}
\alias{simulated_left_bias_heterogeneous_proportions}
\title{Simulated data showing left bias with heterogeneous proportions}
\format{A data frame with 10 rows and 2 variables:
\describe{
 \item{left}{number of parasites on the left side}
 \item{right}{number of parasites on the right side}
 }}
\usage{
simulated_left_bias_heterogeneous_proportions
}
\description{
Simulated data for 10 hosts, providing an example of bilateral asymmetry with left bias
and the left:right ratio varying between hosts.
}
\keyword{datasets}

