% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_symmetrical_infection}
\alias{simulated_symmetrical_infection}
\title{Simulated data showing bilateral symmetry}
\format{A data frame with 10 rows and 2 variables:
\describe{
 \item{left}{number of parasites on the left side}
 \item{right}{number of parasites on the right side}
 }}
\usage{
simulated_symmetrical_infection
}
\description{
Simulated data for 10 hosts, providing an example of bilateral symmetry.
}
\keyword{datasets}

